feat(fetcher): seperate out implementation
move block service implementation to its own package. move traversal helpers to traversal directory
